Move load_mask from _stbt.imgutils -> _stbt.mask
I'm starting to wonder why we need a separate `load_mask` other than
for easy grepping of its callers. I suppose it'll be a good place to
implement a kind of Mask "interning" or caching (or maybe it'll be
enough to cache `load_image`).
