+++ stbt_core/__init__.py
@@ -31,12 +31,13 @@ from _stbt.imgutils import (
     Frame,
     Image,
     load_image,
-    load_mask,
     save_frame)
 from _stbt.keyboard import (
     Keyboard)
 from _stbt.logging import (
     debug)
+from _stbt.mask import (
+    load_mask)
 from _stbt.match import (
     ConfirmMethod,
     match,
